veux_defconfig: Enable power efficient workqueues commit cee22a1 ("workqueues: Introduce new flag WQ_POWER_EFFICIENT for power oriented workqueues") introduced the concept of power efficient workqueues (4 years back), but it was never enabled in upstream kernel configs.
Power efficient workqueues are simply marked as "unbound," so that jobs
queued to them can run on any CPU in the system. It leaves the target
CPU selection to the scheduler, which is the best place for such
decision making. This improves power efficiency for workqueues which are
otherwise pinned to a CPU.

Enable it for ARM64 platforms as ARM platforms were the main target for
the introduction of power efficient workqueues.

Change-Id: I6eefe1e967701c62e557c0db7820c281b1349d08
Signed-off-by: fadlyas07 <mhmmdfdlyas@gmail.com>